Fetch Googletest if can't find it already installed
The developer instructions mention that Googletest needs to be installed
on the user's computer in order to build `chromobius_test`; however,
it's possible that the user has it installed and that CMake doesn't find
it. In addition, the note in the developer docs, about installing
Googletest, is IMHO a bit easy to miss.

We can help both situations by taking advantage of CMake's facilities
for fetching code from a remote git repository. The changes in this
commit add code to `CMakeLists.txt` to try to fetch and build Googletest
if `find_package(GTest QUIET)` does not find it. This makes CMake try a
second time before giving up on being able to build `chromobius_test`.