Set number of `make` jobs based on actions runner
Standard runners on GitHub have 4 vCPUs today, and managed runners and
self-hosted runners may have more. Rather than hardwire the `-j` value
given to `make`, we can set the job number based on the number of
processors available. The change in this commit makes it use 1 less than
the number of processors on the host. (This is currently 3 on the
standard Ubuntu GitHub runners.)