Use apt-get install clang instead of 3rd-party action
In the place where the third-party action `egor/setup-clang` was used,
we can just install clang using `apt`. This is slightly better because
we know the operating system being used (so the portability features of
that third-party action are irrelevant) and it's more future-proof
(because we can't be assured how long the third-party action will be
supported).