Make git not print message about detached heads
Passing the `-c advice.detachedHead=false` option to `git clone` avoids
a message about cloned the repo in a "detached HEAD" state. This is
helpful because when debugging problems, seeing that message in the logs
can was developers' time trying to determine whether it's expected or
whether something has gone wrong.