Put testing into build_sdist and remove check_sdist_installs
The `check_sdist_installs` action didn't do much besides build the sdist
and try to install it, and since `build_sdist` has to do the build too,
we can simply include the testing in `build_sdist`. Doing it this way is
better not just for efficiency: if the test fails, we wouldn't want to
upload the sdist artifact anyway, so this adds a sanity check that was
missing before.