Fix #21 by update and enhancing CMakeLists.txt (#27)
The developer instructions mention that GoogleTest needs to be installed
on the user's computer in order to build `chromobius_test`; however,
they may install it using something like `apt-get install` (which on
some systems only installs the source version and not compiled
libraries) or else they may do something else that counts as installing
it, and yet Chromobius' CMake configuration may still not find it. I
think we can help reduce potential developer confusion and frustration
simply by taking advantage of CMake's facilities for fetching code from
a remote git repository. The GoogleTest documentation [provides
instructions](https://google.github.io/googletest/quickstart-cmake.html#set-up-a-project)
for doing just that. The changes in this commit adds the suggested code
to `CMakeLists.txt` to try to fetch and build GoogleTest if
`find_package(GTest QUIET)` does not find it. This makes CMake try a
second time before giving up on being able to build `chromobius_test`.

In addition, this PR:
* Updates the versions of Stim, Pymatching, and Pybind11 to align with
what's specified in other files
* Adds more status messages so that it's easier to debug where failure
occur
* Fixes some syntax issues involving references to variables
* Slightly reorganizes the CMakeLists.txt file (e.g., to get GoogleTest
earlier rather than at the end),

Note: after applying this PR, there will still be another failure to
find GTest, but this one is coming from Stim. It occurs because
Chromobius' `CMakeLists.txt` fetches and builds Stim and that build
process seems to be isolated in a way that it doesn't find the
GoogleTest that is installed by the earlier steps in our CMakeLists.txt.
(In other words, the changes here make GoogleTest work for target
`chromobius_test` but not for `stim_test`.) This second failure to find
GTest should be solved by updating the versions of Stim this installs.