Revise TestPyPI and PyPI upload jobs
Using PyPI Trusted Publishers avoids the need for using tokens or
setting variables in the workflow and repository, which is good. A
slight complication, though, is that the `pypa/gh-action-pypi-publish`
seems to need `environment` to be defined, which in turn means that if
we want to upload to test.pypi.org as a verification step, we need to do
it using a separate job in order to use different `environment` values.