Set `CHROMOBIUS_VERSION_INFO`
The `CMakeLists.txt file had this in it:

```cmake
add_compile_definitions(CHROMOBIUS_VERSION_INFO=${CHROMOBIUS_VERSION_INFO})
```

However, the variable on the RHS (`CHROMOBIUS_VERSION_INFO`) is not set
when CMake is processing the file.

The new code extracts the version number from the `setup.py` file and
sets the variable before the `add_compile_definitions()` statement.