Update PyPI upload procedure in ci.yml (#35)
The final step of the recently-updated ci.yml (i.e, the step to upload a
release to PyPI) is currently still failing. The process for PyPI
uploads has changed recently, so while updating the workflow, we may as
well take this opportunity to update the upload process to use PyPI's
new "trusted publishers" scheme.

Changes in this PR:

- Use PyPI trusted [publishing
approach](https://docs.pypi.org/trusted-publishers/), which entails
[configuring things on
PyPI](https://docs.pypi.org/trusted-publishers/adding-a-publisher/) and
removing the user name and password from the workflow.

- Add a separate preflight check job using `twine check`. This performs
a fast static check of the metadata in the wheel files. (Although this
could be folded into the test.pypi job, separating it out into its own
job makes it possible to run it independently of the upload job. This is
useful for debugging locally using `act`.)

- Add a separate job to upload to test.pypi.org and then `pip install`
the result, as a final check before uploading to pypi.org. If that
passes, the final job of uploading to pypi.org proceeds.