Fix core dump in `chromobius` by adding `try`-`catch` to `main.cc` (#39)
Currently, if one runs `chromobius` without any arguments, it prints the
help text but also produces a core dump. This happens because the end of
`src/chromobius/commands/main_all.cc` has

```cpp
throw std::invalid_argument(ss.str());
```

and this is not caught by `src/main.cc`. This commit adds a
`try`-`catch` to `src/main.cc` to catch the exception and exit with a
failure code. (This behavior is similar to what `main.perf.cc` does to
exit with an error.)