Fix a few minor issues with CMakeLists.txt (#40)
This does a minor cleanup of a few issues with `CMakeLists.txt`:

* Remove references to variable `PYMATCHING_SOURCE_FILES_NO_MAIN`. This
variable is not defined anywhere in Chromobius and does not exist in
Pymatching, so as far as I can tell, its value is always empty and thus
has no effect in this `CMakeLists.txt` file.

* Set variable `CHROMOBIUS_VERSION_INFO` from `setup.py`. Previously,
the `CMakeLists.txt` file had this in it:
   ```cmake

add_compile_definitions(CHROMOBIUS_VERSION_INFO=${CHROMOBIUS_VERSION_INFO})
   ```
However, the variable on the RHS (`${CHROMOBIUS_VERSION_INFO}`) is not
set when CMake is processing the file. This PR changes it to extract the
version number from the `setup.py` file.

* Wrap long lines and normalize indentation to 4 spaces.