Set `skip-existing` to `true`
I can't think of a good reason to cause the job to fail if an existing
file from the same release is already on pypi. This might happen if you
try to rerun the workflow because of a failure during one of the files.
Let's just let it give a warning and move on.