Update the instructions one more time
Go back to using the ubuntu-24.04 image rather than try to make do with
a smaller one. Also, add a couple more missing items to the Dockerfile
for building a customized local image.

Also, provide the `--no-recurse` flag to the invocation of `act`. That
flag shouldn't be necessary when using `-W`, but if the user starts
experimenting with changing the arguments, showing `--no-recurse` in the
example may help steer them in a direction that leads to fewer
mysterious problems.