Improve `ci.yml`, enhance testability, and document the testing process (#41)
Changes:

* Move more `cibuildwheel` settings into the `pyproject.toml` file,
which both aligns with recommended practices and allows simplifications
to `ci.yml`.

* Get rid of `preflight_check_dists` job in `ci.yml` and make the `twine
check` step be part of the `upload_to_testpypi` job. This simplifies and
streamlines the workflow.

* Get rid of references to `vars.ACT` in `ci.yml`, which turn out to be
unnecessary.

* Enhance local testability by adding support for using a local pypi
server for testing in the `upload_to_testpypi` job.

* Add an input variable `upload_to_pypi` to `workflow_dispatch`
invocations to allow control over whether the upload to `pypi.org` is
done during manual invocations.

* Set `skip-existing` to `true` when using
`actions/gh-action-pypi-publish` so that partial reruns of the workflow
are possible.

* Use explicit versions of runners instead of `-latest`.

* Make miscellaneous small tweaks to `ci.yml`.

* Add a README file to `.github/workflows/` detailing how to do local
testing of `ci.yml`, to document the process.