Need explicitly set pypi user to `__token__` if not passing a value (#43)
The way GitHub Actions workflows treat undefined variables combined with
the way parameters are interpreted by `pypa/gh-action-pypi-publish`
results in that action _not_ using its internal default (which actually
is `__token__`) if the parameter expands to an empty string. My local
testing did not reveal this because in local testing, I always assigned
a value to `testpypi_user`.

Also included in this PR is a clarification in the workflows/README.md
file to one of the debugging tips.