https://test.pypi.org/legacy needs a trailing slash (#44)
It appears to be possible that the current cryptic HTTP 400 error during
upload to test pypi may be due to missing a trailing slash in the URL:
apparently it really _has_ to be `https://test.pypi.org/legacy/` and not
`https://test.pypi.org/legacy`.

This PR also turns off attestations when sending to test.pypi.org, in
case this is another potential source of problems. (Attestations are
still turned on for pypi.org.)