Move link anchors & fix typos & capitalization inconsistencies
With the link anchors on separate lines before section headings, I found
it hard to view and edit the file in an editor. I moved them into the
heading lines, as is done in Stim's documentation; I find this more
readable and easier to edit.

This commit also contains miscellaneous trivial edits to make
capitalization more consistent and fix a couple of typos.