Add description of release process to developer doc (#46)
This adds a section to `doc/developers.md` about making releases of
Chromobius. Much of the text is taken from Stim's developer
documentation.

In addition, this changes the capitalization of some names to make
things more consistent overall, and fixes a couple of typos.

Lastly, it also slightly changes how headings are written in
`doc/developers.md`. The way the link anchors were before, on separate
lines before section headings, I found it hard to view and edit this
file in an editor. By contrast, the approach in the Stim documentation
was easier and more readable (IMHO), so I followed the same pattern
here. I moved the link anchors into the heading lines.